Cache node_features/targets as float16 to buy real disk headroom on Kaggle
Sparse normal storage alone (previous commit) only restored the cache to
its pre-normals size (~12MB/case, ~9.6-10GB total) -- still not enough
margin: a real Kaggle run hit disk exhaustion at 750/800 cases, because
the 9.34GB zip can't be deleted until every case has been extracted from
it (see stream_preprocess_from_zip), so it sits on disk *simultaneously*
with the entire growing cache for the whole preprocessing run. Combined
peak was right at Kaggle's quota.

float16 storage (cast back to float32 immediately on load, so training
math is unaffected -- the project already trains under precision=
"16-mixed" anyway) takes a case to ~9MB, ~7.4GB total: real headroom
against the zip instead of barely refitting the previous size. Verified
locally: max relative rounding error from the cast is ~0.05%, and a
training_step/validation_step smoke test still runs clean.

+++ src/preprocess.py
@@ -33,7 +33,8 @@ def preprocess_case(dataset_root, name, cache_dir, delete_raw_after=False):
         # edge_index as int32 (not PyG's usual int64): node counts are ~180k, well
         # under int32 range, and it's the single biggest piece of the cache -- these
         # two changes take a case from ~24MB to ~12MB, which mattered once a Colab
-        # disk filled up caching the full 800-case split.
+        # disk filled up caching the full 800-case split. Sparse normal storage and
+        # float16 (below) take it down further, to ~9MB -- see their own comments.
         node_features, edge_index, _ = build_graph(simulation)
         surface = simulation.surface
         # normal (columns 5:7 of node_features, see src/graph.py) is exactly
@@ -47,11 +48,22 @@ def preprocess_case(dataset_root, name, cache_dir, delete_raw_after=False):
         targets = np.concatenate(
             [simulation.velocity, simulation.pressure, simulation.nu_t], axis=1
         )
+        # float16 on disk, not float32: halves node_features/targets/normal's
+        # footprint. Cast back to float32 immediately on load
+        # (CachedPyGAirfRANSDataset.get()) so training math is unaffected --
+        # this only shrinks the on-disk representation, and the project
+        # already trains under precision="16-mixed" on Kaggle/Colab anyway,
+        # so it isn't introducing a new precision regime. Needed for real
+        # headroom against the raw zip (9.34GB, undeletable until every case
+        # has been extracted from it -- see stream_preprocess_from_zip)
+        # sitting on disk *simultaneously* with the growing cache for the
+        # entire preprocessing run: this combination hit Kaggle's disk quota
+        # at 750/800 cases even after normals were already stored sparsely.
         payload = {
-            "node_features": torch.tensor(node_features, dtype=torch.float32),
-            "normal_at_surface": torch.tensor(normal_at_surface, dtype=torch.float32),
+            "node_features": torch.tensor(node_features, dtype=torch.float16),
+            "normal_at_surface": torch.tensor(normal_at_surface, dtype=torch.float16),
             "edge_index": torch.tensor(edge_index, dtype=torch.int32),
-            "targets": torch.tensor(targets, dtype=torch.float32),
+            "targets": torch.tensor(targets, dtype=torch.float16),
             "surface": torch.tensor(surface),
             "name": name,
         }
